requirements.txt
Django>=5.0,<6.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0
langchain-tavily>=0.1.0
pinecone-client==3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
